Double Checking Fields are Field

+++ deleteBooking.php
@@ -51,27 +51,29 @@ Bookings table schema:
           $bookingid = $_POST['bookingid'];
           $username = $_SESSION['username'];
 
-          // We determine if the given ID is in the user's bookings
-          $check_sql_id = "SELECT * FROM bookings WHERE b_id = ? AND b_username = ?";
-          $check_stmt_id = mysqli_prepare($conn, $check_sql_id);
-          mysqli_stmt_bind_param($check_stmt_id, "is", $bookingid, $username);
-          mysqli_stmt_execute($check_stmt_id);
-          mysqli_stmt_store_result($check_stmt_id);
-          if (mysqli_stmt_num_rows($check_stmt_id) == 0) {
-              echo "Booking not found. It was either made under a different username or not at all. Please check your bookings for further information.";
-              exit();
-         }
+          if (!empty($bookingid) && !empty($username) {
+            // We determine if the given ID is in the user's bookings
+            $check_sql_id = "SELECT * FROM bookings WHERE b_id = ? AND b_username = ?";
+            $check_stmt_id = mysqli_prepare($conn, $check_sql_id);
+            mysqli_stmt_bind_param($check_stmt_id, "is", $bookingid, $username);
+            mysqli_stmt_execute($check_stmt_id);
+            mysqli_stmt_store_result($check_stmt_id);
+            if (mysqli_stmt_num_rows($check_stmt_id) == 0) {
+                echo "Booking not found. It was either made under a different username or not at all. Please check your bookings for further information.";
+                  exit();
+            }
 
-        // Finally, we delete the booking
-        $sql = "DELETE FROM bookings WHERE b_id = ?";
-        $stmt = mysqli_prepare($conn, $sql);
-        mysqli_stmt_bind_param($stmt, "i", $bookingid);
-        if (mysqli_stmt_execute($stmt)){
-            echo "Booking deleted successfully!";
-            exit();
-        }
-        else {
-            echo "An error has occurred. Please try again.";
+            // Finally, we delete the booking
+            $sql = "DELETE FROM bookings WHERE b_id = ?";
+            $stmt = mysqli_prepare($conn, $sql);
+            mysqli_stmt_bind_param($stmt, "i", $bookingid);
+            if (mysqli_stmt_execute($stmt)){
+                echo "Booking deleted successfully!";
+                exit();
+          }
+          else {
+              echo "An error has occurred. Please try again.";
+          }
         }
       }
     ?>
